applies feedback per review. Next Steps is still TODO
