+++ .snippets/code/tools/typescript-sdk/get-started/terminal-01.html
@@ -1,12 +1,6 @@
 <div id="termynal" data-termynal>
-  <span data-ty="input"
-    ><span class="file-path"></span>npx tsx src/main.ts</span
-  >
+  <span data-ty="input"><span class="file-path"></span>npx tsx src/main.ts</span>
   <span data-ty>Initializing Wormhole SDK...</span>
   <span data-ty>Wormhole SDK Initialized!</span>
-  <span data-ty>Chain ID for Solana Testnet: 1</span>
-  <span data-ty>RPC for Solana Testnet: https://api.devnet.solana.com</span>
-  <span data-ty>Chain ID for Sepolia: 10002</span>
-  <span data-ty>RPC for Sepolia: https://ethereum-sepolia.publicnode.com</span>
   <span data-ty="input"><span class="file-path"></span></span>
 </div>
